playcodesync_demo/requirements.txt
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.1
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.5
daphne==4.0.0
distro==1.7.0
Django==4.1
django-filter==23.3
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.5
more-itertools==8.10.0
msgpack==1.0.7
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
openai==0.28.1
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
SecretStorage==3.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wadllib==1.3.6
yarl==1.9.2
zipp==1.0.0
zope.interface==6.1
